+++ app/src/main/java/com/zhongwang/sale/module/UpdateDataBean.java
@@ -42,11 +42,11 @@ public class UpdateDataBean extends Bean {
 
         private String dmoney;
 
-        private int dbalance;
+        private float dbalance;
 
         private String uname;
 
-        private int dmoney2;
+        private float dmoney2;
 
         private String report_square;
 
@@ -142,7 +142,7 @@ public class UpdateDataBean extends Bean {
             this.dmoney = dmoney;
         }
 
-        public int getDbalance() {
+        public float getDbalance() {
             return dbalance;
         }
 
@@ -158,7 +158,7 @@ public class UpdateDataBean extends Bean {
             this.uname = uname;
         }
 
-        public int getDmoney2() {
+        public float getDmoney2() {
             return dmoney2;
         }
